requirements.txt
asgiref==3.8.1
Django==5.0.4
mysql-connector==2.2.9
sqlparse==0.5.0
python-dotenv==1.0.1
